Read Me
Read Me: A game I programmed long years ago, it is a combination of VB and Multimedia along with audio visual effects.  You may test drive my game by download it and follow simple instructions, it is completely safe to play it on Windows environment. 
Here is the link to download it from:
https://www.dropbox.com/s/wnnvia5nvx1drkv/Asteroids_Buster_r1p4.exe?dl=0